Refactor out FrameGroupID builder
